+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 * Fix - remove ThresholdProfileSet from the domain model
 ## Current Release 
+### 0.114.0 
+**Release Date:** Wed May 23 13:14:24 UTC 2018     
+## Previous Releases 
 ### 0.113.0 
 **Release Date:** Wed May 16 16:41:53 UTC 2018     
-## Previous Releases 
 * Fix - handle lookup not found errors from druid
 ### 0.112.0 
 **Release Date:** Fri May 11 18:47:21 UTC 2018